requirements.txt
Django==4.2.5
djangorestframework==3.14.0
python-dotenv==1.0.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
webcolors==1.11.1
